unify: save some temp work (#2)
